fix: remove unnecessary engine package
