+++ internal/rules/rules_test.go
@@ -3,7 +3,6 @@ package rules
 import (
 	"testing"
 
-	"github.com/Bonial-International-GmbH/sops-compliance-checker/internal/engine"
 	"github.com/Bonial-International-GmbH/sops-compliance-checker/internal/rule"
 	"github.com/stretchr/testify/assert"
 )
@@ -13,21 +12,11 @@ func evalRule(r rule.Rule, trustAnchors ...string) rule.EvalResult {
 	return r.Eval(ctx)
 }
 
-func TestCompile(t *testing.T) {
-	t.Run("compiles fixture", func(t *testing.T) {
-		compiled, err := Compile(configFixture.Rules)
-		assert.NoError(t, err)
-		assert.Equal(t, rulesFixture, compiled)
-	})
-}
-
 func TestNestedRules(t *testing.T) {
-	engine := engine.New(rulesFixture)
+	rootRule := rulesFixture
 
-	t.Run("all trust anchors missing", func(t *testing.T) {
-		trustAnchors := []string{}
-
-		result := engine.Eval(trustAnchors)
+	t.Run("no trust anchors", func(t *testing.T) {
+		result := evalRule(rootRule)
 		assert.False(t, result.Success)
 		assert.Len(t, result.Matched.Slice(), 0)
 		assert.Len(t, result.Unmatched.Slice(), 0)
@@ -42,7 +31,7 @@ func TestNestedRules(t *testing.T) {
 			"arn:aws:kms:eu-west-1:123456789012:alias/production-cicd",
 		}
 
-		result := engine.Eval(trustAnchors)
+		result := evalRule(rootRule, trustAnchors...)
 		assert.True(t, result.Success)
 		assert.Len(t, result.Matched.Slice(), len(trustAnchors))
 		assert.Len(t, result.Unmatched.Slice(), 0)
@@ -58,7 +47,7 @@ func TestNestedRules(t *testing.T) {
 			"i don't belong here",
 		}
 
-		result := engine.Eval(trustAnchors)
+		result := evalRule(rootRule, trustAnchors...)
 		assert.True(t, result.Success)
 		assert.Len(t, result.Matched.Slice(), len(trustAnchors)-1)
 		assert.Equal(t, []string{"i don't belong here"}, result.Unmatched.Slice())
